+++ src/pages/Layout/Header/index.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Headers } from './header.style'
+
+const Header = () => {
+  return (
+    <Headers
+      style={{
+        padding: '0.8rem',
+      }}
+    ></Headers>
+  )
+}
+
+export default Header
