+++ src/pages/Other/index.js
@@ -1,20 +1,7 @@
 import React from 'react'
-import { useNavigate } from 'react-router-dom'
 
 const Other = () => {
-  const navigate = useNavigate()
-
-  const logout = () => {
-    sessionStorage.removeItem('user')
-    navigate('/login')
-  }
-
-  return (
-    <div>
-      Other
-      <button onClick={logout}> Logout</button>
-    </div>
-  )
+  return <div>Other</div>
 }
 
 export default Other
